Improve matching of strings with parentheses and non-string arguments.
Previously, strings with parentheses weren't being matched. It was also
matching the function declarations like "function $t(...)". So use more
regex-fu to more strictly match only strings, trying to account for
quoting, etc. This is probably a bad idea to be doing this type of thing
with regexes, but it seems to work...